+++ src/Components/Utils/RecipeParts.js
@@ -3,6 +3,14 @@ import TextField from "@mui/material/TextField";
 import Typography from "@mui/material/Typography";
 import Paper from "@mui/material/Paper";
 import Box from "@mui/material/Box";
+import Chip from "@mui/material/Chip";
+import Autocomplete from "@mui/material/Autocomplete";
+import StarIcon from "@mui/icons-material/Star";
+import Stack from "@mui/material/Stack";
+
+import FavoriteSwitch from "../Utils/FavoriteSwitch";
+
+import { createKey, updateRequest } from "../../utils/requests";
 
 export const renderEditingButtons = (onCancel, onSave) => (
   <>
@@ -59,5 +67,119 @@ export const renderNotesInput = (notes, updateNotes) => (
       updateNotes(event.target.value);
     }}
     variant="standard"
+  />
+);
+
+const renderFavorite = (editable, isFavorite, updateIsFavorite) => {
+  if (editable) {
+    return (
+      <FavoriteSwitch isChecked={isFavorite} updateChecked={updateIsFavorite} />
+    );
+  }
+
+  if (isFavorite) {
+    return (
+      <Chip
+        key={"favorite"}
+        label={
+          <StarIcon
+            sx={{
+              "&&": {
+                color: "alt.main",
+                verticalAlign: "bottom",
+              },
+            }}
+            fontSize="small"
+          />
+        }
+        size="small"
+        variant="outlined"
+        color="tertiary"
+      />
+    );
+  }
+
+  return null;
+};
+
+export const renderTagList = (
+  editable,
+  { tags, isFavorite },
+  updateIsFavorite,
+  getDeleteTagHandler,
+  glossaryRecipeTags
+) => (
+  <Stack
+    direction="row"
+    spacing={1}
+    key="tags"
+    sx={{ width: "95%" }}
+    alignItems={"center"}
+  >
+    {renderFavorite(editable, isFavorite, updateIsFavorite)}
+    {tags.map((tagId) => (
+      <Chip
+        key={tagId}
+        label={<Typography>{glossaryRecipeTags[tagId]}</Typography>}
+        size="small"
+        variant="outlined"
+        color="tertiary"
+        onDelete={editable ? getDeleteTagHandler(tagId) : undefined}
+      />
+    ))}
+  </Stack>
+);
+
+export const renderTagControls = (
+  tags,
+  addTag,
+  glossaryRecipeTags,
+  glossaryPath
+) => (
+  <Autocomplete
+    id={"addtagSelect"}
+    options={
+      glossaryRecipeTags
+        ? Object.keys(glossaryRecipeTags).map((tagId) => ({
+            tagId,
+            title: glossaryRecipeTags[tagId],
+          }))
+        : []
+    }
+    getOptionLabel={(option) => option.title}
+    getOptionDisabled={(option) => tags.includes(option.tagId)}
+    filterOptions={(options, params) => {
+      const { inputValue, getOptionLabel } = params;
+      const filtered = options.filter((option) =>
+        getOptionLabel(option)
+          .toLocaleUpperCase()
+          .includes(inputValue.toUpperCase())
+      );
+      const isExisting = options.some((option) => inputValue === option.title);
+      if (inputValue !== "" && !isExisting) {
+        filtered.push({
+          inputValue,
+          title: `Create "${inputValue}"`,
+        });
+      }
+      return filtered;
+    }}
+    value={null}
+    onChange={(event, selectedOption) => {
+      const { tagId: _tagId, inputValue } = selectedOption;
+      let tagId = _tagId;
+      if (inputValue) {
+        tagId = createKey(`${glossaryPath}/recipeTags`);
+        updateRequest({
+          [`${glossaryPath}/recipeTags/${tagId}`]: inputValue,
+        });
+      }
+      addTag(tagId);
+    }}
+    renderInput={(params) => (
+      <TextField {...params} label="Enter tag" size="small" />
+    )}
+    blurOnSelect={true}
+    clearOnBlur={true}
   />
 );
